Rename Component::Text::ClipboardData to Client::System::Input::ClipBoard
